api/requirements.txt
fastapi[all]
fastapi-offline
uvicorn
python-multipart
motor
pycryptodome